Refactor PostgreSQL database adapter and core storage modules
- Simplified core DatabaseAdapter abstract class by removing interface methods
- Updated Agent class to handle optional database registration
- Removed redundant database and type files from examples directory

This commit streamlines the database storage architecture and enhances modularity across the Binkai ecosystem.
